updated capi.rs to expose more functionality, added grug.h to match capi.rs, fixed issue in build.rs

+++ gruggers/build.rs
@@ -15,32 +15,41 @@ fn build_tests() {
 	let archive_path = var("OUT_DIR").unwrap();
 
 	println!("cargo::rustc-link-search={}", out_dir.display());
-	println!("cargo::rerun-if-changed={}", test_source_path);
-	println!("cargo::rerun-if-changed={}", bench_source_path);
 
 	#[cfg(target_os = "linux")]
 	{
-		_ = std::fs::copy(test_source_path + "/libtests.so", archive_path.clone() + "/libtests.so");
-		_ = std::fs::copy(bench_source_path + "/libbench.so", archive_path + "/libbench.so");
+		if Ok(true) = std::fs::exists(&test_source_path) {
+			println!("cargo::rerun-if-changed={}", test_source_path);
+			_ = std::fs::copy(test_source_path + "/libtests.so", archive_path.clone() + "/libtests.so");
+		}
+		if Ok(true) = std::fs::exists(&bench_source_path) {
+			println!("cargo::rerun-if-changed={}", bench_source_path);
+			_ = std::fs::copy(bench_source_path + "/libbench.so", archive_path + "/libbench.so");
+		}
 	}
 	#[cfg(target_os = "windows")]
 	{
-		let tests_dll_out_path = archive_path.clone() + "/tests.dll";
-		for optional_path in ["tests.dll"] {
-			_ = std::fs::copy(test_source_path.clone() + optional_path, &tests_dll_out_path);
+		if let Ok(true) = std::fs::exists(&test_source_path) {
+			println!("cargo::rerun-if-changed={}", test_source_path);
+			let tests_dll_out_path = archive_path.clone() + "/tests.dll";
+			for optional_path in ["tests.dll"] {
+				_ = std::fs::copy(test_source_path.clone() + optional_path, &tests_dll_out_path);
+			}
+			let tests_lib_out_path = archive_path.clone() + "/tests.lib";
+			for optional_path in ["tests.lib", "libtests.dll.a"] {
+				_ = std::fs::copy(test_source_path.clone() + optional_path, &tests_lib_out_path);
+			}
 		}
-		let tests_lib_out_path = archive_path.clone() + "/tests.lib";
-		for optional_path in ["tests.lib", "libtests.dll.a"] {
-			_ = std::fs::copy(test_source_path.clone() + optional_path, &tests_lib_out_path);
-		}
-
-		let bench_dll_out_path = archive_path.clone() + "/bench.dll";
-		for optional_path in ["bench.dll", "libbench.dll"] {
-			_ = std::fs::copy(bench_source_path.clone() + optional_path, &bench_dll_out_path);
-		}
-		let bench_lib_out_path = archive_path + "/bench.lib";
-		for optional_path in ["bench.lib", "libbench.lib"] {
-			_ = std::fs::copy(bench_source_path.clone() + optional_path, &bench_lib_out_path);
+		if let Ok(true) = std::fs::exists(&bench_source_path) {
+			println!("cargo::rerun-if-changed={}", bench_source_path);
+			let bench_dll_out_path = archive_path.clone() + "/bench.dll";
+			for optional_path in ["bench.dll", "libbench.dll"] {
+				_ = std::fs::copy(bench_source_path.clone() + optional_path, &bench_dll_out_path);
+			}
+			let bench_lib_out_path = archive_path + "/bench.lib";
+			for optional_path in ["bench.lib", "libbench.lib"] {
+				_ = std::fs::copy(bench_source_path.clone() + optional_path, &bench_lib_out_path);
+			}
 		}
 	}
 }
